arm64: Add workaround for Cavium erratum 27456
On ThunderX T88 pass 2.x parts, TLBI does not cause the icache to
be changed based on the asid.  This can cause random crashes and
wrong result as the wrong code is being executed.

This patch implements the workaround (which flushes the local icache
when switching the mm) by using code patching.

Signed-off-by: Andrew Pinski <apinski@cavium.com>
Signed-off-by: David Daney <ddaney@caviumnetworks.com>