all: s390: make compat wrappers the generic solution
The problem that makes us to use wrappers is that some compat
architectures allows user code to access top halves of registers.
This is not a problem for syscalls that are already handled by compat
code, or for that who has types that are the same size in kernel and
userspace. In case of lp64/ilp32 the problem is in pointer types, long,
unsigned long. S390 folks already have a solution for it. In this patch,
it is turned to be general, as arm64/ilp32 needs it too.

Build-tested on s390.